feat: added subcommands to aw-sync, support selectively syncing buckets
