+++ aw-sync/src/lib.rs
@@ -7,3 +7,6 @@ extern crate serde_json;
 mod sync;
 pub use sync::sync_datastores;
 pub use sync::sync_run;
+
+mod accessmethod;
+pub use accessmethod::AccessMethod;
