+++ aw-sync/src/main.rs
@@ -16,10 +16,11 @@ extern crate serde_json;
 use std::path::Path;
 
 use chrono::{DateTime, Utc};
-use clap::Parser;
+use clap::{Parser, Subcommand};
 
 use aw_client_rust::AwClient;
 
+mod accessmethod;
 mod sync;
 
 const DEFAULT_PORT: &str = "5600";
@@ -27,6 +28,9 @@ const DEFAULT_PORT: &str = "5600";
 #[derive(Parser)]
 #[clap(version = "0.1", author = "Erik Bjäreholt")]
 struct Opts {
+    #[clap(subcommand)]
+    command: Commands,
+
     /// Host of instance to connect to
     #[clap(long, default_value = "127.0.0.1")]
     host: String,
@@ -40,16 +44,26 @@ struct Opts {
     /// If not specified, exit
     #[clap(long)]
     sync_dir: String,
-    /// Date to start syncing from
-    /// If not specified, start from beginning
-    /// NOTE: might be unstable, as count cannot be used to verify integrity of sync.
-    /// Format: YYYY-MM-DD
-    #[clap(long)]
-    start_date: Option<String>,
-    /// Specify buckets to sync
-    /// If not specified, all buckets will be synced
-    #[clap(long)]
-    buckets: Vec<String>,
+}
+
+#[derive(Subcommand)]
+enum Commands {
+    /// Clones repos
+    #[clap(arg_required_else_help = true)]
+    Sync {
+        /// Date to start syncing from
+        /// If not specified, start from beginning
+        /// NOTE: might be unstable, as count cannot be used to verify integrity of sync.
+        /// Format: YYYY-MM-DD
+        #[clap(long)]
+        start_date: Option<String>,
+        /// Specify buckets to sync
+        /// If not specified, all buckets will be synced
+        #[clap(long)]
+        buckets: Vec<String>,
+    },
+    /// List buckets and their sync status
+    List {},
 }
 
 fn main() -> std::io::Result<()> {
@@ -74,16 +88,37 @@ fn main() -> std::io::Result<()> {
 
     let client = AwClient::new(opts.host.as_str(), port, "aw-sync-rust");
 
-    let start: Option<DateTime<Utc>> = opts.start_date.map(|date| {
-        chrono::DateTime::parse_from_rfc3339(&date)
-            .unwrap()
-            .with_timezone(&chrono::Utc)
-    });
-    sync::sync_run(sync_directory, client, opts.buckets, start).map_err(|e| {
+    match &opts.command {
+        // Perform two-way sync
+        Commands::Sync {
+            start_date,
+            buckets,
+        } => {
+            let start: Option<DateTime<Utc>> = start_date.as_ref().map(|date| {
+                let date_copy = date.clone();
+                chrono::DateTime::parse_from_rfc3339(&date_copy)
+                    .unwrap()
+                    .with_timezone(&chrono::Utc)
+            });
+
+            sync::sync_run(sync_directory, client, buckets, start).map_err(|e| {
+                println!("Error: {}", e);
+                std::io::Error::new(std::io::ErrorKind::Other, e)
+            })?;
+            info!("Finished successfully, exiting...");
+            Ok(())
+        }
+
+        // List all buckets
+        Commands::List {} => {
+            sync::list_buckets(&client, sync_directory);
+            Ok(())
+        }
+    }
+    .map_err(|e: String| {
         println!("Error: {}", e);
         std::io::Error::new(std::io::ErrorKind::Other, e)
     })?;
-    info!("Finished successfully, exiting...");
 
     // Needed to give the datastores some time to commit before program is shut down.
     // 100ms isn't actually needed, seemed to work fine with as little as 10ms, but I'd rather give
