+++ aw-sync/src/sync.rs
@@ -9,124 +9,19 @@ extern crate chrono;
 extern crate reqwest;
 extern crate serde_json;
 
-use std::collections::HashMap;
 use std::fs;
 use std::path::{Path, PathBuf};
 
 use aw_client_rust::AwClient;
 use chrono::{DateTime, Duration, Utc};
-use reqwest::StatusCode;
 
 use aw_datastore::{Datastore, DatastoreError};
 use aw_models::{Bucket, Event};
 
-// This trait should be implemented by both AwClient and Datastore, unifying them under a single API
-pub trait AccessMethod: std::fmt::Debug {
-    fn get_buckets(&self) -> Result<HashMap<String, Bucket>, String>;
-    fn get_bucket(&self, bucket_id: &str) -> Result<Bucket, DatastoreError>;
-    fn create_bucket(&self, bucket: &Bucket) -> Result<(), DatastoreError>;
-    fn get_events(
-        &self,
-        bucket_id: &str,
-        start: Option<DateTime<Utc>>,
-        end: Option<DateTime<Utc>>,
-        limit: Option<u64>,
-    ) -> Result<Vec<Event>, String>;
-    fn insert_events(&self, bucket_id: &str, events: Vec<Event>) -> Result<Vec<Event>, String>;
-    fn get_event_count(&self, bucket_id: &str) -> Result<i64, String>;
-    fn heartbeat(&self, bucket_id: &str, event: Event, duration: f64) -> Result<(), String>;
-}
+use crate::accessmethod::AccessMethod;
 
-impl AccessMethod for Datastore {
-    fn get_buckets(&self) -> Result<HashMap<String, Bucket>, String> {
-        Ok(self.get_buckets().unwrap())
-    }
-    fn get_bucket(&self, bucket_id: &str) -> Result<Bucket, DatastoreError> {
-        self.get_bucket(bucket_id)
-    }
-    fn create_bucket(&self, bucket: &Bucket) -> Result<(), DatastoreError> {
-        self.create_bucket(bucket)?;
-        self.force_commit().unwrap();
-        Ok(())
-    }
-    fn get_events(
-        &self,
-        bucket_id: &str,
-        start: Option<DateTime<Utc>>,
-        end: Option<DateTime<Utc>>,
-        limit: Option<u64>,
-    ) -> Result<Vec<Event>, String> {
-        Ok(self.get_events(bucket_id, start, end, limit).unwrap())
-    }
-    fn heartbeat(&self, bucket_id: &str, event: Event, duration: f64) -> Result<(), String> {
-        self.heartbeat(bucket_id, event, duration).unwrap();
-        self.force_commit().unwrap();
-        Ok(())
-    }
-    fn insert_events(&self, bucket_id: &str, events: Vec<Event>) -> Result<Vec<Event>, String> {
-        let res = self.insert_events(bucket_id, &events[..]).unwrap();
-        self.force_commit().unwrap();
-        Ok(res)
-    }
-    fn get_event_count(&self, bucket_id: &str) -> Result<i64, String> {
-        Ok(self.get_event_count(bucket_id, None, None).unwrap())
-    }
-}
-
-impl AccessMethod for AwClient {
-    fn get_buckets(&self) -> Result<HashMap<String, Bucket>, String> {
-        Ok(self.get_buckets().unwrap())
-    }
-    fn get_bucket(&self, bucket_id: &str) -> Result<Bucket, DatastoreError> {
-        let bucket = self.get_bucket(bucket_id);
-        match bucket {
-            Ok(bucket) => Ok(bucket),
-            Err(e) => {
-                warn!("{:?}", e);
-                let code = e.status().unwrap();
-                if code == StatusCode::NOT_FOUND {
-                    Err(DatastoreError::NoSuchBucket(bucket_id.into()))
-                } else {
-                    panic!("Unexpected error");
-                }
-            }
-        }
-    }
-    fn get_events(
-        &self,
-        bucket_id: &str,
-        start: Option<DateTime<Utc>>,
-        end: Option<DateTime<Utc>>,
-        limit: Option<u64>,
-    ) -> Result<Vec<Event>, String> {
-        Ok(self.get_events(bucket_id, start, end, limit).unwrap())
-    }
-    fn insert_events(&self, _bucket_id: &str, _events: Vec<Event>) -> Result<Vec<Event>, String> {
-        //Ok(self.insert_events(bucket_id, &events[..]).unwrap())
-        Err("Not implemented".to_string())
-    }
-    fn get_event_count(&self, bucket_id: &str) -> Result<i64, String> {
-        Ok(self.get_event_count(bucket_id).unwrap())
-    }
-    fn create_bucket(&self, bucket: &Bucket) -> Result<(), DatastoreError> {
-        self.create_bucket(bucket.id.as_str(), bucket._type.as_str())
-            .unwrap();
-        Ok(())
-        //Err(DatastoreError::InternalError("Not implemented".to_string()))
-    }
-    fn heartbeat(&self, bucket_id: &str, event: Event, duration: f64) -> Result<(), String> {
-        self.heartbeat(bucket_id, &event, duration)
-            .map_err(|e| format!("{:?}", e))
-    }
-}
-
-/// Performs a single sync pass
-pub fn sync_run(
-    sync_directory: &Path,
-    client: AwClient,
-    buckets: Vec<String>,
-    start: Option<DateTime<Utc>>,
-) -> Result<(), String> {
+fn setup_local_remote(client: &AwClient, sync_directory: &Path) -> Result<Datastore, String> {
+    // FIXME: Don't run twice if already exists
     fs::create_dir_all(sync_directory).unwrap();
 
     let info = client.get_info().unwrap();
@@ -142,9 +37,22 @@ pub fn sync_run(
     let ds_localremote = Datastore::new(dbfile, false);
     info!("Set up remote for local device");
 
+    Ok(ds_localremote)
+}
+
+/// Performs a single sync pass
+pub fn sync_run(
+    sync_directory: &Path,
+    client: AwClient,
+    buckets: &Vec<String>,
+    start: Option<DateTime<Utc>>,
+) -> Result<(), String> {
+    let ds_localremote = setup_local_remote(&client, sync_directory)?;
+
     //let ds_remotes = setup_test(sync_directory).unwrap();
     //info!("Set up remotes for testing");
 
+    let info = client.get_info().unwrap();
     let remote_dbfiles = find_remotes_nonlocal(sync_directory, info.device_id.as_str());
     info!("Found remotes: {:?}", remote_dbfiles);
 
@@ -154,7 +62,7 @@ pub fn sync_run(
     // Pull
     info!("Pulling...");
     for ds_from in &ds_remotes {
-        sync_datastores(ds_from, &client, false, None);
+        sync_datastores(ds_from, &client, false, None, &buckets);
     }
 
     // Push local server buckets to sync folder
@@ -164,15 +72,29 @@ pub fn sync_run(
         &ds_localremote,
         true,
         Some(info.device_id.as_str()),
+        &buckets,
     );
 
-    log_buckets(&client);
+    list_buckets(&client, sync_directory);
+
+    Ok(())
+}
+
+pub fn list_buckets(client: &AwClient, sync_directory: &Path) {
+    let ds_localremote = setup_local_remote(client, sync_directory).unwrap();
+
+    let info = client.get_info().unwrap();
+    let remote_dbfiles = find_remotes_nonlocal(sync_directory, info.device_id.as_str());
+    info!("Found remotes: {:?}", remote_dbfiles);
+
+    // TODO: Check for compatible remote db version before opening
+    let ds_remotes: Vec<Datastore> = remote_dbfiles.iter().map(create_datastore).collect();
+
+    log_buckets(client);
     log_buckets(&ds_localremote);
     for ds_from in &ds_remotes {
         log_buckets(ds_from);
     }
-
-    Ok(())
 }
 
 /// Returns a list of all remote dbs
@@ -182,7 +104,7 @@ fn find_remotes(sync_directory: &Path) -> std::io::Result<Vec<PathBuf>> {
         .map(|res| res.ok().unwrap().path())
         .filter(|p| p.is_dir())
         .flat_map(|d| {
-            println!("{}", d.to_str().unwrap());
+            //println!("{}", d.to_str().unwrap());
             fs::read_dir(d).unwrap()
         })
         .map(|res| res.ok().unwrap().path())
@@ -316,6 +238,7 @@ pub fn sync_datastores(
     ds_to: &dyn AccessMethod,
     is_push: bool,
     src_did: Option<&str>,
+    buckets: &Vec<String>,
 ) {
     // FIXME: "-synced" should only be appended when synced to the local database, not to the
     // staging area for local buckets.
@@ -333,6 +256,8 @@ pub fn sync_datastores(
             }
             tup.1.clone()
         })
+        // Filter out buckets not in the buckets vec
+        .filter(|bucket| buckets.iter().any(|b_id| b_id == &bucket.id))
         .collect();
 
     // Sync buckets in order of most recently updated
